Refactor ConversionUtils methods to PathUtils and AssetUtils
