+++ src/toniarts/openkeeper/game/controller/GameController.java
@@ -74,7 +74,6 @@ import toniarts.openkeeper.game.trigger.door.DoorTriggerLogicController;
 import toniarts.openkeeper.game.trigger.object.ObjectTriggerLogicController;
 import toniarts.openkeeper.game.trigger.party.PartyTriggerLogicController;
 import toniarts.openkeeper.game.trigger.player.PlayerTriggerLogicController;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.map.KeeperSpell;
 import toniarts.openkeeper.tools.convert.map.KwdFile;
 import toniarts.openkeeper.tools.convert.map.Player;
@@ -198,7 +197,7 @@ public class GameController implements IGameLogicUpdatable, AutoCloseable, IGame
             if (level != null) {
 
                 kwdFile = new KwdFile(Main.getDkIIFolder(),
-                        Paths.get(ConversionUtils.getRealFileName(Main.getDkIIFolder(), PathUtils.DKII_MAPS_FOLDER + level + ".kwd")));
+                        Paths.get(PathUtils.getRealFileName(Main.getDkIIFolder(), PathUtils.DKII_MAPS_FOLDER + level + ".kwd")));
 
             } else {
                 kwdFile.load();
