+++ src/toniarts/openkeeper/game/state/GameState.java
@@ -47,7 +47,6 @@ import toniarts.openkeeper.game.trigger.creature.CreatureTriggerLogicController;
 import toniarts.openkeeper.game.trigger.door.DoorTriggerLogicController;
 import toniarts.openkeeper.game.trigger.object.ObjectTriggerLogicController;
 import toniarts.openkeeper.game.trigger.party.PartyTriggerLogicController;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.map.KeeperSpell;
 import toniarts.openkeeper.tools.convert.map.KwdFile;
 import toniarts.openkeeper.tools.convert.map.Player;
@@ -158,7 +157,7 @@ public class GameState extends AbstractPauseAwareState implements IGameLogicUpda
                     // Load the level data
                     if (level != null) {
                         kwdFile = new KwdFile(Main.getDkIIFolder(),
-                                Paths.get(ConversionUtils.getRealFileName(Main.getDkIIFolder() + PathUtils.DKII_MAPS_FOLDER, level + ".kwd")));
+                                Paths.get(PathUtils.getRealFileName(Main.getDkIIFolder() + PathUtils.DKII_MAPS_FOLDER, level + ".kwd")));
                     } else {
                         kwdFile.load();
                     }
