+++ src/toniarts/openkeeper/gui/Cursor.java
@@ -26,7 +26,7 @@ import java.io.File;
 import java.nio.ByteBuffer;
 import java.nio.IntBuffer;
 import toniarts.openkeeper.tools.convert.AssetsConverter;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
+import toniarts.openkeeper.utils.PathUtils;
 
 public class Cursor extends JmeCursor {
 
@@ -58,7 +58,7 @@ public class Cursor extends JmeCursor {
             throw new IllegalArgumentException("The cursor needs at least a framecount of 1.");
         }
 
-        Texture tex = assetManager.loadTexture(ConversionUtils.convertFileSeparators(AssetsConverter.MOUSE_CURSORS_FOLDER.concat(File.separator).concat(Filename)));
+        Texture tex = assetManager.loadTexture(PathUtils.convertFileSeparators(AssetsConverter.MOUSE_CURSORS_FOLDER.concat(File.separator).concat(Filename)));
         Image img = tex.getImage();
         // width must be a multiple of 16, otherwise the cursor gets distorted
         int width = img.getWidth() % 16 == 0 ? img.getWidth() : (img.getWidth() - img.getWidth() % 16) + 16;
