+++ src/toniarts/openkeeper/tools/convert/KmfModelLoader.java
@@ -70,6 +70,7 @@ import toniarts.openkeeper.tools.convert.kmf.MeshVertex;
 import toniarts.openkeeper.tools.convert.kmf.Triangle;
 import toniarts.openkeeper.tools.convert.kmf.Uv;
 import toniarts.openkeeper.tools.modelviewer.ModelViewer;
+import toniarts.openkeeper.utils.AssetUtils;
 import toniarts.openkeeper.utils.PathUtils;
 
 /**
@@ -638,7 +639,7 @@ public class KmfModelLoader implements AssetLoader {
                 } else {
 
                     // Ok, it it not in the cache yet, but maybe it has been already generated, so use it and update the defaults in it
-                    fileName = ConversionUtils.stripFileName(mat.getName());
+                    fileName = PathUtils.stripFileName(mat.getName());
 
                     // If there are multiple texture options, add a suffix to the material file name
                     if (mat.getTextures().size() > 1) {
@@ -759,7 +760,7 @@ public class KmfModelLoader implements AssetLoader {
     private Texture loadTexture(String texture, AssetInfo assetInfo) {
 
         // Load the texture
-        TextureKey textureKey = new TextureKey(ConversionUtils.getCanonicalAssetKey(AssetsConverter.TEXTURES_FOLDER.concat("/").concat(texture).concat(".png")), false);
+        TextureKey textureKey = new TextureKey(AssetUtils.getCanonicalAssetKey(AssetsConverter.TEXTURES_FOLDER.concat("/").concat(texture).concat(".png")), false);
         Texture tex = assetInfo.getManager().loadTexture(textureKey);
         return tex;
     }
