+++ src/toniarts/openkeeper/tools/convert/conversion/task/ConvertHiScores.java
@@ -22,9 +22,9 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import toniarts.openkeeper.game.data.HiScores;
 import toniarts.openkeeper.tools.convert.AssetsConverter;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.hiscores.HiScoresEntry;
 import toniarts.openkeeper.tools.convert.hiscores.HiScoresFile;
+import toniarts.openkeeper.utils.PathUtils;
 
 /**
  * Dungeon Keeper II hiscores conversion. Converts all hiscore files to our own
@@ -56,7 +56,7 @@ public class ConvertHiScores extends ConversionTask {
         try {
 
             // Load the original
-            Path file = Paths.get(ConversionUtils.getRealFileName(dungeonKeeperFolder, "Data/Settings/HiScores.dat"));
+            Path file = Paths.get(PathUtils.getRealFileName(dungeonKeeperFolder, "Data/Settings/HiScores.dat"));
             HiScoresFile originalHiScores = new HiScoresFile(file);
 
             // Convert it!
