+++ src/toniarts/openkeeper/tools/convert/conversion/task/ConvertMapThumbnails.java
@@ -31,7 +31,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.imageio.ImageIO;
 import toniarts.openkeeper.tools.convert.AssetsConverter;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.map.GameLevel;
 import toniarts.openkeeper.tools.convert.map.KwdFile;
 import toniarts.openkeeper.utils.AssetUtils;
@@ -124,7 +123,7 @@ public class ConvertMapThumbnails extends ConversionTask {
         // TODO maybe image size in Settings ???
         BufferedImage thumbnail = MapThumbnailGenerator.generateMap(kwd, 144, 144, false);
 
-        Path destinationPath = Paths.get(destination, ConversionUtils.stripFileName(kwd.getGameLevel().getName()) + ".png");
+        Path destinationPath = Paths.get(destination, PathUtils.stripFileName(kwd.getGameLevel().getName()) + ".png");
         try (OutputStream os = Files.newOutputStream(destinationPath, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
                 BufferedOutputStream bos = new BufferedOutputStream(os)) {
             ImageIO.write(thumbnail, "png", bos);
