+++ src/toniarts/openkeeper/tools/convert/conversion/task/ConvertModels.java
@@ -37,7 +37,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import toniarts.openkeeper.tools.convert.AssetsConverter;
 import static toniarts.openkeeper.tools.convert.AssetsConverter.getAssetsFolder;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.KmfAssetInfo;
 import toniarts.openkeeper.tools.convert.KmfModelLoader;
 import toniarts.openkeeper.tools.convert.kmf.KmfFile;
@@ -118,7 +117,7 @@ public class ConvertModels extends ConversionTask {
         // Meshes are in the data folder, access the packed file
         WadFile wad;
         try {
-            wad = new WadFile(Paths.get(ConversionUtils.getRealFileName(dungeonKeeperFolder + PathUtils.DKII_DATA_FOLDER, "Meshes.WAD")));
+            wad = new WadFile(Paths.get(PathUtils.getRealFileName(dungeonKeeperFolder + PathUtils.DKII_DATA_FOLDER, "Meshes.WAD")));
         } catch (IOException ex) {
             throw new RuntimeException("Could not open the meshes.wad archive!", ex);
         }
