+++ src/toniarts/openkeeper/tools/convert/conversion/task/ConvertSounds.java
@@ -28,7 +28,6 @@ import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import toniarts.openkeeper.tools.convert.AssetsConverter;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.sound.SdtFile;
 import toniarts.openkeeper.utils.AssetUtils;
 import toniarts.openkeeper.utils.PathUtils;
@@ -68,7 +67,7 @@ public class ConvertSounds extends ConversionTask {
         final List<Path> sdtFiles = new ArrayList<>();
         Path dataDir = null;
         try {
-            dataDir = Paths.get(ConversionUtils.getRealFileName(dungeonKeeperFolder, dataDirectory));
+            dataDir = Paths.get(PathUtils.getRealFileName(dungeonKeeperFolder, dataDirectory));
             Files.walkFileTree(dataDir, new SimpleFileVisitor<Path>() {
                 @Override
                 public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
