+++ src/toniarts/openkeeper/view/map/Water.java
@@ -39,7 +39,6 @@ import java.util.Map;
 import java.util.logging.Logger;
 import toniarts.openkeeper.common.EntityInstance;
 import toniarts.openkeeper.tools.convert.AssetsConverter;
-import toniarts.openkeeper.tools.convert.ConversionUtils;
 import toniarts.openkeeper.tools.convert.map.ArtResource;
 import toniarts.openkeeper.tools.convert.map.Terrain;
 import toniarts.openkeeper.utils.AssetUtils;
@@ -110,7 +109,7 @@ public class Water {
             // Lava
             mat = new Material(assetManager,
                     "Common/MatDefs/Light/Lighting.j3md");
-            TextureKey textureKey = new TextureKey(ConversionUtils.getCanonicalAssetKey(AssetsConverter.TEXTURES_FOLDER.concat("/").concat(entityInstances.get(0).getEntity().getTopResource().getName()).concat(".png")), false);
+            TextureKey textureKey = new TextureKey(AssetUtils.getCanonicalAssetKey(AssetsConverter.TEXTURES_FOLDER.concat("/").concat(entityInstances.get(0).getEntity().getTopResource().getName()).concat(".png")), false);
             Texture tex = assetManager.loadTexture(textureKey);
             mat.setTexture("DiffuseMap", tex);
             MapViewController.setTerrainMaterialLighting(mat, entityInstances.get(0).getEntity());
